finished repositories & services? minus the token and external api stuffs?

+++ program/Cryptocop.Software.API.Repositories/Implementations/AddressRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Cryptocop.Software.API.Models.Dtos;
+using Cryptocop.Software.API.Models.Entities;
 using Cryptocop.Software.API.Models.InputModels;
 using Cryptocop.Software.API.Repositories.Interfaces;
 using Cryptocop.Software.API.Repositories.Data;
@@ -12,12 +13,30 @@ public class AddressRepository : IAddressRepository
     private readonly CryptocopDbContext _context;
     public AddressRepository(CryptocopDbContext context) => _context = context;
 
-
-    public Task AddAddressAsync(string email, AddressInputModel address)
+    // Create
+    public async Task AddAddressAsync(string email, AddressInputModel address)
     {
-        throw new NotImplementedException();
+        var user = await _context.Users.FirstOrDefaultAsync(u => u.Email == email);
+        if (user == null) throw new ArgumentException($"User with email: {email} not found");
+
+        var newAddress = new Address
+        {
+            UserId = user.Id,
+            StreetName = address.StreetName,
+            HouseNumber = address.HouseNumber,
+            ZipCode = address.ZipCode,
+            Country = address.Country,
+            City = address.City,
+        };
+
+        user.Addresses.Add(newAddress);
+        _context.Addresses.Add(newAddress);
+
+        _context.SaveChanges();
+        return;
     }
 
+    // Get all
     public async Task<IEnumerable<AddressDto>> GetAllAddressesAsync(string email)
     {
         var user = await _context.Users.FirstOrDefaultAsync(u => u.Email == email);
@@ -28,24 +47,17 @@ public class AddressRepository : IAddressRepository
         return ret.Select(o => o.ToDto());
     }
 
-    // helper func, takes in users list of addresses
-    // public async Task<AddressDto> GetAddressAsync(string email, int id)
-    // {
-    //     var addresses = await GetAllAddressesAsync(email);
-    //     if (addresses == null) throw new ArgumentException("User has no saved addresses");
-
-    //     var addr = addresses.FirstOrDefault(a => a.Id == id);
-    //     if (addr == null) throw new ArgumentException($"No address found with ID {id}");
-
-    //     return addr;
-    // }
-
-    public Task DeleteAddressAsync(string email, int addressId)
+    // Delete
+    public async Task DeleteAddressAsync(string email, int addressId)
     {
+        var user = await _context.Users.FirstOrDefaultAsync(u => u.Email == email);
+        if (user == null) throw new ArgumentException($"User with email: {email} not found");
+        
         var addrs = _context.Addresses.FirstOrDefault(a => a.Id == addressId);
         if (addrs == null) throw new ArgumentException($"No address found with ID {addressId}");
 
+        user.Addresses.Remove(addrs);
         _context.Addresses.Remove(addrs);
-        return Task.CompletedTask;
+        return;
     }
 }
